Mostly working now TODO: not show social dialogs for other users TODO: not subscribe to message tap on both
