+++ src/SSW.Rewards/MauiProgram.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Logging;
 using Mopups.Hosting;
 using SkiaSharp.Views.Maui.Controls.Hosting;
 using SSW.Rewards.Mobile.Controls;
+using SSW.Rewards.Mobile.ViewModels.ProfileViewModels;
 using System.Reflection;
 using ZXing.Net.Maui.Controls;
 using IBrowser = IdentityModel.OidcClient.Browser.IBrowser;
@@ -42,15 +43,23 @@ public static class MauiProgram
 
         // TODO: move this to a source genertor
 
-        var profileViewModel = typeof(ProfileViewModel);
+        var excludedTypes = new Type[]
+        {
+            typeof(OtherProfileViewModel),
+            typeof(OthersProfilePage),
+            typeof(ProfileViewModelBase),
+        };
 
         var definedTypes = Assembly.GetExecutingAssembly().DefinedTypes
-            .Where(e => e.IsSubclassOf(typeof(Page)) || e.IsSubclassOf(typeof(BaseViewModel)));
+            .Where(e => e.IsSubclassOf(typeof(Page)) || e.IsSubclassOf(typeof(BaseViewModel)) && !excludedTypes.Contains(e));
 
         foreach (var type in definedTypes)
         {
             builder.Services.AddSingleton(type.AsType());
         }
+
+        builder.Services.AddTransient<OtherProfileViewModel>();
+        builder.Services.AddTransient<OthersProfilePage>();
 
         builder.Services.AddSingleton<ILeaderService, LeaderService>();
         builder.Services.AddSingleton<IUserService, UserService>();
