+++ src/SSW.Rewards/Pages/MyProfilePage.xaml
@@ -1,0 +1,83 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             BackgroundColor="{StaticResource MainBackground}"
+             xmlns:controls="clr-namespace:SSW.Rewards.Mobile.Controls"
+             xmlns:helpers="clr-namespace:SSW.Rewards.Mobile.Helpers"
+             xmlns:viewModels="clr-namespace:SSW.Rewards.Mobile.ViewModels.ProfileViewModels"
+             xmlns:mct="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
+             x:DataType="viewModels:MyProfileViewModel"
+             x:Class="SSW.Rewards.Mobile.Pages.MyProfilePage">
+    <ContentPage.Behaviors>
+        <mct:StatusBarBehavior StatusBarColor="{StaticResource MainBackground}" StatusBarStyle="LightContent"/>
+    </ContentPage.Behaviors>
+    <ContentPage.Resources>
+        <DataTemplate x:Key="AchievementsTemplate">
+            <controls:AchievementGrid />
+        </DataTemplate>
+
+        <DataTemplate x:Key="NotificationsTemplate">
+            <controls:Notifications />
+        </DataTemplate>
+
+        <DataTemplate x:Key="ActivityTemplate">
+            <controls:RecentActivity />
+        </DataTemplate>
+
+        <helpers:ProfileCarouselSelector x:Key="ProfileViewSelector"
+                                         Notifications="{StaticResource NotificationsTemplate}"
+                                         Achievements="{StaticResource AchievementsTemplate}"
+                                         RecentActivity="{StaticResource ActivityTemplate}" />
+    </ContentPage.Resources>
+
+    <!-- TECH DEBT: The following workaround adds an extra row on iOS.
+                    This is necessary due to a bug in .NET MAUI Preview that 
+                    causes part of the page to be hidden behind the tab bar.
+                    See: https://github.com/dotnet/maui/issues/17817
+    -->
+    <Grid RowDefinitions="{OnPlatform 'Auto,Auto,*,10', iOS='Auto,Auto,*,10,80'}"
+          RowSpacing="6"
+          Padding="20,20,20,5">
+
+        <Border
+            Grid.Row="0"
+            HorizontalOptions="End"
+            BackgroundColor="{StaticResource MainBackground}"
+            StrokeShape="Ellipse"
+            StrokeThickness="0"
+            Padding="0"
+            IsVisible="{Binding ShowPopButton}">
+            <Border.GestureRecognizers>
+                <TapGestureRecognizer Command="{Binding PopProfile}"/>
+            </Border.GestureRecognizers>
+            <Label
+                FontFamily="FluentIcons"
+                Text="&#xe4c3;"
+                FontSize="35"
+                TextColor="{StaticResource primary}"/>
+        </Border>
+
+        <controls:ProfileStats Grid.Row="1"/>
+
+        <CarouselView Grid.Row="2"
+                      Margin="0,20,0,0"
+                      IndicatorView="ProfileIndicator"
+                      ItemsSource="{Binding ProfileSections}"
+                      ItemTemplate="{StaticResource ProfileViewSelector}"/>
+
+        <IndicatorView Grid.Row="3"
+                       SelectedIndicatorColor="{StaticResource primary}"
+                       IndicatorColor="{StaticResource IndicatorColor}"
+                       HorizontalOptions="Center"
+                       x:Name="ProfileIndicator"/>
+
+        <ActivityIndicator Grid.Row="0"
+                           Grid.RowSpan="4"
+                           HorizontalOptions="Center"
+                           VerticalOptions="Center"
+                           IsRunning="True"
+                           Color="{StaticResource primary}"
+                           IsVisible="{Binding IsLoading}"/>
+
+    </Grid>
+</ContentPage>
