+++ src/SSW.Rewards/Pages/ProfilePages/OthersProfilePage.xaml.cs
@@ -1,0 +1,35 @@
+﻿using SSW.Rewards.Mobile.ViewModels.ProfileViewModels;
+
+namespace SSW.Rewards.Mobile.Pages;
+
+public partial class OthersProfilePage : ContentPage
+{
+    private bool _initialised;
+
+    private OtherProfileViewModel viewModel;
+
+    public OthersProfilePage(OtherProfileViewModel vm, LeaderViewModel leader)
+    {
+        InitializeComponent();
+        viewModel = vm;
+
+        viewModel.SetLeader(leader);
+
+        viewModel.Navigation = Navigation;
+        BindingContext = viewModel;
+    }
+
+    protected override async void OnAppearing()
+    {
+        if (!_initialised)
+            await viewModel.Initialise();
+
+        _initialised = true;
+    }
+
+    protected override void OnDisappearing()
+    {
+        base.OnDisappearing();
+        viewModel.OnDisappearing();
+    }
+}
