--- src/SSW.Rewards/ViewModels/OtherProfileViewModel.cs
@@ -1,8 +1,0 @@
-﻿namespace SSW.Rewards.Mobile.ViewModels;
-
-public class OtherProfileViewModel : ProfileViewModel
-{
-    public OtherProfileViewModel(IRewardService rewardsService, IUserService userService, ISnackbarService snackbarService) : base(rewardsService, userService, snackbarService)
-    {
-    }
-}
